Bugfix: Return value of App\Entity\RequestForComment::getYesShare() must be of the type integer, float returned

+++ src/Entity/RequestForComment.php
@@ -182,7 +182,7 @@ class RequestForComment
             $total++;
         }
 
-        return round($yes / $total * 100, 0);
+        return (int) round($yes / $total * 100, 0);
     }
 
     public function getCurrentResults()
